working strategy

+++ regulation/path_regulation_gr14.h
@@ -29,7 +29,7 @@ void follow_path(CtrlStruct *cvs, vector<array<float,2> > path);
  *			l_speed	point on left speed
  *			r_speed point on right speed
  */
-void get_new_speed(float gamma, float *l_speed, float *r_speed);
+void get_new_speed(double gamma, double *l_speed, double *r_speed);
 
 NAMESPACE_CLOSE();
